Add files via upload
The Micro Scale Biogas Project is located in Nepal and the purpose of the project activity is the set up 1 (one) independent biogas plant (digester) of 20m3 capacity for serving three (3) households using polutry litter and food waste. Co-digestion, as is the case in this project activity, leads to enhanced biogas production potential, system stability due to synergetic effects, and resolving the problem of waste management in the vicinity in a holistic approach.